+++ tests/Pest.php
@@ -1,0 +1,5 @@
+<?php
+
+use DilovanMatini\Enumable\Tests\TestCase;
+
+uses(TestCase::class)->in(__DIR__);