+++ tests/TestCase.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace DilovanMatini\Enumable\Tests;
+
+use Orchestra\Testbench\TestCase as Orchestra;
+
+abstract class TestCase extends Orchestra
+{
+    protected function getPackageProviders($app): array
+    {
+        return [];
+    }
+}